requirements.txt
python-dotenv==1.0.1
sqlalchemy==2.0.28
boto3==1.34.69
psycopg2==2.9.9
pandas==2.2.1
requests==2.31.0